Fixed so you can define fd before redirection operator and it stucks to the operator if there is no space in between
