+++ inc/tokenization.h
@@ -1,29 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minishell.h                                        :+:      :+:    :+:   */
+/*   tokenization.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: emajuri <emajuri@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/04/03 15:00:45 by emajuri           #+#    #+#             */
-/*   Updated: 2023/04/13 16:30:43 by emajuri          ###   ########.fr       */
+/*   Created: 2023/04/13 16:18:06 by emajuri           #+#    #+#             */
+/*   Updated: 2023/04/13 16:20:58 by emajuri          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef MINISHELL_H
-# define MINISHELL_H
 
-# define METACHARS "|<> \t\n"
+#ifndef TOKENIZATION_H
+# define TOKENIZATION_H
 
-# include "../libft/libft.h"
-# include <stdio.h>
-# include <readline/readline.h>
-# include <readline/history.h>
-# include <stdlib.h>
-# include <unistd.h>
-
-void	rl_replace_line(const char *text, int clear_undo);
-void	print_error(const char *error_message, char *input);
-char	**tokenization(char *pipeline);
+int	ft_isspace(char c);
+int	len_word(char *str);
+int	len_metachars(char *str, int metachar);
+int	len_delim_word(char *str, int delim);
 
 #endif
